python-ai-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
python-multipart==0.0.9

langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38
langchain-openai==0.1.6

pypdf==4.1.0

tiktoken

pydantic<2